+++ src/components/top-bar/top-bar-button.tsx
@@ -1,4 +1,4 @@
-import { IconButton, IconButtonProps } from "@chakra-ui/react";
+import { Circle, Float, IconButton, IconButtonProps } from "@chakra-ui/react";
 import { IconProp } from "@fortawesome/fontawesome-svg-core";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { Tooltip } from "../ui/tooltip";
@@ -7,8 +7,14 @@ function TopBarButton({
   tooltip,
   icon,
   trigger,
+  notification,
   ...rest
-}: IconButtonProps & { tooltip: string; icon: IconProp; trigger?: string }) {
+}: IconButtonProps & {
+  tooltip: string;
+  icon: IconProp;
+  trigger?: string;
+  notification?: boolean;
+}) {
   return (
     <Tooltip
       lazyMount
@@ -23,10 +29,16 @@ function TopBarButton({
       <IconButton
         size={"sm"}
         variant={"ghost"}
+        rounded="full"
         {...rest}
         color={{ base: "gray.700", _dark: "gray.300" }}
       >
         <FontAwesomeIcon icon={icon} />
+        {notification && (
+          <Float placement="top-end" offsetX="2" offsetY="2">
+            <Circle bg="red" size="10px" />
+          </Float>
+        )}
       </IconButton>
     </Tooltip>
   );